fixed redux tests
