+++ client/store/breakfast.spec.js
@@ -1,12 +1,7 @@
 /* global describe beforeEach afterEach it */
 
 import {expect} from 'chai'
-import {
-  getAllBreakfasts,
-  getSingleBreakfast,
-  increaseQuantity,
-  decreaseQuantity
-} from './breakfast'
+import {getAllBreakfasts, getSingleBreakfast} from './breakfast'
 import axios from 'axios'
 import MockAdapter from 'axios-mock-adapter'
 import configureMockStore from 'redux-mock-store'
@@ -53,32 +48,6 @@ describe('thunk creators', () => {
       await store.dispatch(getSingleBreakfast(1))
       const actions = store.getActions()
       expect(actions[0].type).to.be.equal('GOT_SINGLE_BREAKFAST')
-      expect(actions[0].breakfast).to.be.deep.equal(fakeBreakfast)
-    })
-  })
-
-  describe('increaseQuantity', () => {
-    it('eventually dispatches the UPDATED QUANTITY action', async () => {
-      const fakeBreakfast = {name: 'Cheerios', quantity: 1}
-      mockAxios
-        .onPut('/api/breakfasts/1/increase')
-        .replyOnce(201, fakeBreakfast)
-      await store.dispatch(increaseQuantity(1))
-      const actions = store.getActions()
-      expect(actions[0].type).to.be.equal('UPDATED_QUANTITY')
-      expect(actions[0].breakfast).to.be.deep.equal(fakeBreakfast)
-    })
-  })
-
-  describe('decreaseQuantity', () => {
-    it('eventually dispatches the UPDATED QUANTITY action', async () => {
-      const fakeBreakfast = {name: 'Cheerios', quantity: 0}
-      mockAxios
-        .onPut('/api/breakfasts/1/decrease')
-        .replyOnce(201, fakeBreakfast)
-      await store.dispatch(decreaseQuantity(1))
-      const actions = store.getActions()
-      expect(actions[0].type).to.be.equal('UPDATED_QUANTITY')
       expect(actions[0].breakfast).to.be.deep.equal(fakeBreakfast)
     })
   })
